+++ graphbrain/learner/classifier.py
@@ -3,7 +3,7 @@ import logging
 
 from graphbrain.hyperedge import hedge
 from graphbrain.learner.pattern_ops import (common_pattern, is_variable, contains_variable, remove_variables,
-                                            all_variables, merge_patterns)
+                                            all_variables, merge_patterns, apply_variables)
 from graphbrain.learner.rule import Rule
 from graphbrain.learner.rule import from_json as rule_from_json
 from graphbrain.patterns import is_wildcard
@@ -64,35 +64,6 @@ def apply_curly_brackets(edge):
         return edge
     else:
         return hedge([apply_curly_brackets(subedge) for subedge in edge])
-
-
-def apply_variable(edge, var_name, var_edge):
-    if edge == var_edge:
-        return hedge(('var', edge, var_name)), True
-
-    subedges = []
-    found = False
-    if edge.not_atom:
-        for subedge in edge:
-            if found:
-                subedges.append(subedge)
-            else:
-                vedge, result = apply_variable(subedge, var_name, var_edge)
-                subedges.append(vedge)
-                if result:
-                    found = True
-        return hedge(subedges), found
-
-    return edge, False
-
-
-def apply_variables(edge, variables):
-    new_edge = edge
-    for var_name, var_edge in variables.items():
-        new_edge, found = apply_variable(new_edge, var_name, var_edge)
-        if not found:
-            return None
-    return new_edge
 
 
 def _atom_mappings2atom_fun(atom_mappings):